feat(prototyper): support RV32 and Yuzuki Neko (#273)
Make startup, relocation, trap handling, PMP configuration and CSR
access account for XLEN. Support RV32 timer and PMU operations while
preserving RV64 support.

Add the Yuzuki Neko platform configuration with firmware loaded at
0x40f80000 and an S-mode jump target of 0x40000000.

Adapt test-kernel for RV32 and verify it on both qemu-system-riscv32
and qemu-system-riscv64.

+++ firmware/prototyper/src/platform/qemu_aplic/registers.rs
@@ -49,9 +49,9 @@ impl EncodedMsiAddress {
             return Err(Error::InvalidArgs);
         }
 
-        let base_ppn = base.as_usize() >> PAGE_SHIFT;
-        if base_ppn & low_bits_mask(hart_index_bits) != 0
-            || base_ppn >> u32::BITS > low_bits_mask(HIGH_PPN_WIDTH)
+        let base_ppn = (base.as_usize() as u64) >> PAGE_SHIFT;
+        if base_ppn & low_bits_mask(hart_index_bits) as u64 != 0
+            || base_ppn >> u32::BITS > low_bits_mask(HIGH_PPN_WIDTH) as u64
         {
             return Err(Error::InvalidArgs);
         }
